# chore(parser): Parser error optimisation