--- a/crates/noirc_frontend/src/parser/errors.rs
+++ b/crates/noirc_frontend/src/parser/errors.rs
@@ -67,8 +67,8 @@ impl ParserError {
 
 impl std::fmt::Display for ParserError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut expected = vecmap(self.expected_tokens.iter(), ToString::to_string);
-        expected.append(&mut vecmap(self.expected_labels.iter(), |label| format!("{label}")));
+        let mut expected = vecmap(&self.expected_tokens, ToString::to_string);
+        expected.append(&mut vecmap(&self.expected_labels, |label| format!("{label}")));
 
         if expected.is_empty() {
             write!(f, "Unexpected {} in input", self.found)