--- a/crates/noirc_frontend/src/parser/labels.rs
+++ b/crates/noirc_frontend/src/parser/labels.rs
@@ -2,8 +2,10 @@ use std::fmt;
 
 use crate::token::TokenKind;
 
+/// Used to annotate parsing rules with extra context that can be presented to the user later in
+/// the case of an error.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ParserLabel {
+pub enum ParsingRuleLabel {
     Atom,
     BinaryOperator,
     Cast,
@@ -19,22 +21,22 @@ pub enum ParserLabel {
     TokenKind(TokenKind),
 }
 
-impl fmt::Display for ParserLabel {
+impl fmt::Display for ParsingRuleLabel {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
-            ParserLabel::Atom => write!(f, "atom"),
-            ParserLabel::BinaryOperator => write!(f, "binary operator"),
-            ParserLabel::Cast => write!(f, "cast"),
-            ParserLabel::Expression => write!(f, "expression"),
-            ParserLabel::FieldAccess => write!(f, "field access"),
-            ParserLabel::Global => write!(f, "global"),
-            ParserLabel::IntegerType => write!(f, "integer type"),
-            ParserLabel::Parameter => write!(f, "parameter"),
-            ParserLabel::Pattern => write!(f, "pattern"),
-            ParserLabel::Statement => write!(f, "statement"),
-            ParserLabel::Term => write!(f, "term"),
-            ParserLabel::TypeExpression => write!(f, "type expression"),
-            ParserLabel::TokenKind(token_kind) => write!(f, "{:?}", token_kind),
+            ParsingRuleLabel::Atom => write!(f, "atom"),
+            ParsingRuleLabel::BinaryOperator => write!(f, "binary operator"),
+            ParsingRuleLabel::Cast => write!(f, "cast"),
+            ParsingRuleLabel::Expression => write!(f, "expression"),
+            ParsingRuleLabel::FieldAccess => write!(f, "field access"),
+            ParsingRuleLabel::Global => write!(f, "global"),
+            ParsingRuleLabel::IntegerType => write!(f, "integer type"),
+            ParsingRuleLabel::Parameter => write!(f, "parameter"),
+            ParsingRuleLabel::Pattern => write!(f, "pattern"),
+            ParsingRuleLabel::Statement => write!(f, "statement"),
+            ParsingRuleLabel::Term => write!(f, "term"),
+            ParsingRuleLabel::TypeExpression => write!(f, "type expression"),
+            ParsingRuleLabel::TokenKind(token_kind) => write!(f, "{:?}", token_kind),
         }
     }
 }