… crates/noirc_frontend/src/parser/parser.rs