# Related issue(s)

<!-- If it does not already exist, first create a GitHub issue that describes the problem this Pull Request (PR) solves before creating the PR and link it here. -->

Resolves # <!-- link to issue -->

# Description

## Summary of changes

tl;dr
~40% speed up in release mode
~10% speed up in debug mode

Chumsky produces and aggregates large numbers of errors in normal operation regardless of whether it's on the happy path. This is because as it tries each grammar rule in turn, it produces an error until the matching rule is encountered. Currently our error construction and aggregation is rather expensive, largely in due to the construction and destruction of `BTreeSet`s that may hold zero or very few items.

This PR uses `SmallOrdSet` to avoid heap allocations as long as possible.

Other changes:
- labels switched to an enum to avoid string allocations

Samples taken using `time nargo print-acir`

|              | debug  | release |
|--------|--------|---------|
| master | 4.1s      |     0.47s |
| this branch | 3.7s |  0.34s |

## Dependency additions / changes

<!-- If applicable. -->

## Test additions / changes

<!-- If applicable. -->

# Checklist

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/) and/or `cargo fmt` with default settings.
- [ ] I have [linked](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue) this PR to the issue(s) that it resolves.
- [ ] I have reviewed the changes on GitHub, line by line.
- [ ] I have ensured all changes are covered in the description.

## Documentation needs
- [ ] This PR requires documentation updates when merged.

<!-- If checked, list / describe what needs to be documented. -->

# Additional context

<!-- If applicable. -->