@joss-aztec 2023-05-05
Just noticed these changes throw up the following clippy warning:

```
warning: the `Err`-variant returned from this function is very large
   --> crates/noirc_frontend/src/ast/mod.rs:154:10
    |
154 |     ) -> Result<UnresolvedTypeExpression, ParserError> {
    |          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the `Err`-variant is at least 344 bytes
    |
    = help: try reducing the size of `parser::errors::ParserError`, for example by boxing large elements or replacing it with `Box<parser::errors::ParserError>`
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#result_large_err
    = note: `#[warn(clippy::result_large_err)]` on by default
```

Guess I better benchmark `Resolver::resolve_expression` to check I haven't inadvertently hurt it. If not, then I guess it's justifiable to mute the warning.

I'm aware that I've spent more time on this PR than it's really worth - so shout if you think I should abandon.