@jfecher 2023-05-05
I think this PR is still a worthwhile pursuit, its nice to have the timing information so that we know for later what any performance characteristics are like. I'm interested in what the performance of tinyset would be as well? Is it an easy slot-in replacement for small-ord-set?