@joss-aztec 2023-05-09
I had a go at `tinyset`, but it's pretty cumbersome to integrate with our current enums. You need to implement the following trait on your element type:
```rs
pub trait Fits64: Copy {
    /// Convert back *from* a u64.  This is unsafe, since it is only
    /// infallible (and lossless) if the `u64` originally came from
    /// type `Self`.
    unsafe fn from_u64(x: u64) -> Self;
    /// Convert to a `u64`.  This should be infallible.
    fn to_u64(self) -> u64;
}
```
Probably the biggest hurdle is you'd need to map to some alternative representation of a token that's able to support the `Copy` trait. (Maybe an argument to split data from ids as we have done for ssa?) I also had trouble with `unsafe` related errors, but that could be my rust inexperience rather than a real blocker.